Se agregan Excepciones y clases BeneficioDescuentoPorcentaje y Beneficio2x1

+++ ejercicio-tarjeta2/src/main/java/ar/edu/untref/aydoo/Producto.java
@@ -3,14 +3,14 @@ package ar.edu.untref.aydoo;
 public class Producto {
 
 	private Beneficio beneficioProducto;
-	private int precioProducto;
+	private double precioProducto;
 
-	public Producto(int precio, Beneficio beneficio){
+	public Producto(double precio, Beneficio beneficio){
 		this.precioProducto = precio;
 		this.beneficioProducto = beneficio;
 	}
 	
-	public int getPrecio(){
+	public double getPrecio(){
 		return this.precioProducto;
 	}
